requirements.txt
# LiveKit and AI Agent dependencies
livekit-agents[google]>=1.2.0
livekit-plugins-google
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-noise-cancellation~=0.2
livekit
livekit-api
google-generativeai
google-genai>=1.0.0
google-api-core
faster-whisper
soundfile
pyaudio
pyttsx3==2.90
speechrecognition==3.8.1
numpy>=1.24.0,<2.0.0
scipy
flask
python-dotenv
langchain-community
duckduckgo-search
requests
uvicorn
requests
uvicorn
flask-websockets